Calculate open quarter, add to selection options
