+++ src/filing/App.jsx
@@ -12,6 +12,8 @@ import isRedirecting from './actions/isRedirecting.js'
 import updateFilingPeriod from './actions/updateFilingPeriod.js'
 import { detect } from 'detect-browser'
 import { FilingAnnouncement } from './common/FilingAnnouncement'
+import { splitYearQuarter } from './api/utils.js'
+import { currentQuarterlyPeriod } from './constants/dates'
 
 import 'normalize.css'
 import './app.css'
@@ -77,6 +79,13 @@ export class AppContainer extends Component {
     return !!props.location.pathname.match(/^\/filing\/\d{4}\/$/)
   }
 
+  isValidPeriod(period) {
+    const [year] = splitYearQuarter(period)
+    const filingPeriods = this.props.config.filingPeriods
+
+    return filingPeriods.indexOf(year) !== -1 || period === currentQuarterlyPeriod()
+  }
+
   render() {
     const { match: { params }, location, config: { filingPeriods, filingAnnouncement } } = this.props
 
@@ -89,7 +98,7 @@ export class AppContainer extends Component {
         <ConfirmationModal />
         {isBeta() ? <Beta/> : null}
         {filingAnnouncement ? <FilingAnnouncement data={filingAnnouncement} /> : null}
-        {filingPeriods.indexOf(params.filingPeriod.split('-')[0]) !== -1
+        {this.isValidPeriod(params.filingPeriod)
           ? this._renderAppContents(this.props)
           : params.filingPeriod === '2017'
             ? <p className="full-width">Files are no longer being accepted for the 2017 filing period. For further assistance, please contact <a href="mailto:hmdahelp@cfpb.gov">HMDA Help</a>.</p>
